+++ lua/plugins/config/focus_config.lua
@@ -4,7 +4,7 @@ return function(use)
 		config = function()
 			require('focus').setup({
 				signcolumn = false, -- disable sign column only on main window
-				compatible_filetrees = { 'nvimtree', 'nerdtree', 'chadtree', 'fern', 'Calltree' },
+				excluded_filetypes = { 'calltree' },
 			})
 		end,
 	})